+++ cmd/neofs-node/config.go
@@ -90,9 +90,10 @@ type applicationConfiguration struct {
 	}
 
 	engine struct {
-		errorThreshold uint32
-		shardPoolSize  uint32
-		shards         []storage.ShardCfg
+		errorThreshold         uint32
+		shardPoolSize          uint32
+		shards                 []storage.ShardCfg
+		isIgnoreUninitedShards bool
 	}
 
 	policer struct {
@@ -154,6 +155,7 @@ func (a *applicationConfiguration) readConfig(c *config.Config) error {
 
 	a.engine.errorThreshold = engineconfig.ShardErrorThreshold(c)
 	a.engine.shardPoolSize = engineconfig.ShardPoolSize(c)
+	a.engine.isIgnoreUninitedShards = engineconfig.IgnoreUninitedShards(c)
 
 	// Morph